converter/requirements.txt
fastapi
uvicorn[standard]
lxml
jsonschema
xmltodict
python-multipart
pdfminer.six
pytest
pdfplumber
ebooklib
beautifulsoup4
pandas
python-docx
odfpy
markdownify
PyMuPDF